+++ Changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### Version 1.16.5 - 1.0.5
+ - added the `/additionalredstone` command
+    - `/additionalredstone github` shows the url to the github page
+    - `/additionalredstone discord` shows the url to the discord server
+    - `/additionalredstone wiki` shows the url to the wiki
+    - `/additionalredstone issues` shows the url to the issues page
+    - `/additionalredstone curseforge` shows the url to the curseforge page
+    - `/additionalredstone modrinth` shows the url to the modrinth page
+
 ### Version 1.16.5 - 1.0.2
  - add sequencer recipe to the circuit-maker